vuejs/requirements.txt
fastapi==0.70.0
uvicorn
python-multipart
jinja2
pydantic
python-dotenv
